+++ afisha/admin.py
@@ -2,10 +2,11 @@ from django.contrib.admin import register
 from django.contrib import admin
 
 from afisha.models import Event, EventParticipant
+from users.utils import StaffRequiredAdminMixin
 
 
 @register(Event)
-class EventAdmin(admin.ModelAdmin):
+class EventAdmin(StaffRequiredAdminMixin, admin.ModelAdmin):
     list_display = (
         "address",
         "contact",
@@ -19,8 +20,16 @@ class EventAdmin(admin.ModelAdmin):
     )
     empty_value_display = "-пусто-"
 
+    def get_queryset(self, request):
+        if request.user.profile.is_moderator_reg:
+            return Event.objects.filter(city=request.user.profile.city)
+        return Event.objects.all()
+
+    def has_view_permission(self, request, obj=None):
+        return request.user.profile.is_moderator_reg or request.user.profile.is_moderator_gen
+
 
 @register(EventParticipant)
-class EventParticipantAdmin(admin.ModelAdmin):
+class EventParticipantAdmin(StaffRequiredAdminMixin, admin.ModelAdmin):
     list_display = ("user", "event")
     empty_value_display = "-пусто-"
